requirements.txt
python-keystoneclient
openstacksdk